Separate Request screens to their files
